requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
Flask-WTF==1.2.1
Flask-Limiter==3.5.1
gunicorn==21.2.0
Werkzeug==2.3.6